更新：调整包结构，解决循环引用
