+++ consul/_init/init.go
@@ -1,10 +1,11 @@
-package consul
+package _init
 
 import (
 	"github.com/kataras/iris/v12"
 	"github.com/kataras/iris/v12/mvc"
 	libConfig "github.com/lm1996-mojor/go-core-library/config"
 	_const "github.com/lm1996-mojor/go-core-library/const"
+	"github.com/lm1996-mojor/go-core-library/consul"
 	"github.com/lm1996-mojor/go-core-library/global"
 	"github.com/lm1996-mojor/go-core-library/log"
 	"github.com/lm1996-mojor/go-core-library/store"
@@ -12,7 +13,7 @@ import (
 )
 
 func Init(app *iris.Application) {
-	mvc.New(app.Party("/consul")).Handle(NewController())
+	mvc.New(app.Party("/consul")).Handle(consul.NewController())
 	log.Info("初始化服务治理-服务健康检查接口")
 }
 
@@ -29,18 +30,18 @@ func init() {
 		}
 		searchConditionValue := "ID contains " + libConfig.Sysconfig.App.Name + "-" + host + ":" + libConfig.Sysconfig.App.Port
 
-		consulServiceInfoList, err := FindServiceList(searchConditionValue)
+		consulServiceInfoList, err := consul.FindServiceList(searchConditionValue)
 		if err != nil {
 			panic("查询服务列表失败：" + err.Error())
 		}
 		if len(consulServiceInfoList) > 0 {
 			for serviceId, _ := range consulServiceInfoList {
-				if err1 := ServiceDeregister(serviceId); err1 != nil {
+				if err1 := consul.ServiceDeregister(serviceId); err1 != nil {
 					panic("服务注销失败：" + err1.Error())
 				}
 			}
 		}
-		consulServiceId := Register()
+		consulServiceId := consul.Register()
 		store.Set(_const.ConsulEndId, consulServiceId)
 		global.RegisterInit(global.Initiator{Action: Init, Level: runLevel, EndFlag: false})
 	}
